requirements.txt
# Core Flask dependencies
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.7

# Environment and configuration
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.21
psycopg2-binary==2.9.7

# Security
bcrypt==4.0.1

# QR Code generation
qrcode==7.4.2
Pillow==10.0.1

# HTTP requests for Telegram
requests==2.31.0

# Excel export
openpyxl==3.1.2

# Production server (optional - for deployment with Gunicorn)
gunicorn==21.2.0

# Additional production dependencies
email-validator==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
blinker==1.6.3